Implemented date range selection for transactions
The current default display is one month before and after today's date.
The spinners are set to the correct dates, but the default view is
showing transactions from the start date until the present date instead
of from the start date to the defined end date (one month in the
future).  If the date picker is selected and done is pressed, the
transaction list displays the expected range. Need to try to talk
through this with Daniel.